Testing scripts

+++ src/main/java/data/Alignment.java
@@ -65,7 +65,7 @@ public class Alignment {
     String s = "Alignment type: " + type + "\n" +
         "Sequence length: " + sequenceLength + "\n" +
         "Graph size: " + graphSize + "\n" +
-        "Score: " + score + "\n" +
+        "Score: " + (int) score + "\n" +
         "Time: " + time + "\n" +
         "Alignment: ";
     if (alignment != null) {
